update(cache 0.5)

+++ apps/articles/views.py
@@ -6,6 +6,8 @@ from articles.forms import ArticleEditForm
 
 from django.core.paginator import Paginator
 
+from django.core.cache import cache
+
 
 def list_views(request):
     articles = Article.objects.annotate(like_count=Count('like'))
@@ -18,38 +20,51 @@ def list_views(request):
 
 
 def detail_views(request, pk):
-
     article = Article.objects.get(pk=pk)
-    like_count = len(article.like_set.all())
 
-    self_like = Like.objects.filter(user_id=request.user.id, article_id=pk).exists()
-    if self_like:
-        button_label = 'Unlike'
-    else:
-        button_label = 'Like'
+    like_count = cache.get('like_count')
+    if not like_count:
 
-    self_favorite = Favorite.objects.filter(user_id=request.user.id, article_id=pk).exists()
-    if self_favorite:
-        f_button_label = 'Unfavorite'
-    else:
-        f_button_label = 'Favorite'
+        like_count = len(article.like_set.all())
+        cache.set('like_count', like_count)
+
+    self_like = cache.get('self_like')
+    if not self_like:
+
+        self_like = Like.objects.filter(user_id=request.user.id, article_id=pk).exists()
+        if self_like:
+            button_label = 'Unlike'
+        else:
+            button_label = 'Like'
+
+    self_favorite = cache.get('self_favorite')
+    if not self_favorite:
+
+        self_favorite = Favorite.objects.filter(user_id=request.user.id, article_id=pk).exists()
+        if self_favorite:
+            f_button_label = 'Unfavorite'
+        else:
+            f_button_label = 'Favorite'
 
     return render(request,'article.html', {'article':article, 'like_count':like_count, 'button_label':button_label, 'f_button_label':f_button_label})
 
-def like_articles_views(request):
 
+def like_articles_views(request):
     user_id = request.POST.get('user_id')
     article_id = request.POST.get('article_id')
 
-    self_like = Like.objects.filter(user_id=user_id, article_id=article_id).exists()
-    if self_like:
-        article_like = Like.objects.filter(user_id=user_id, article_id=article_id)
-        article_like.delete()
-        button_label = 'Like'
-    else:
-        article_like = Like(user_id=user_id, article_id=article_id)
-        article_like.save()
-        button_label = 'Unlike'
+    self_like = cache.get('self_like')
+    if not self_like:
+
+        self_like = Like.objects.filter(user_id=user_id, article_id=article_id).exists()
+        if self_like:
+            article_like = Like.objects.filter(user_id=user_id, article_id=article_id)
+            article_like.delete()
+            button_label = 'Like'
+        else:
+            article_like = Like(user_id=user_id, article_id=article_id)
+            article_like.save()
+            button_label = 'Unlike'
 
     like_count = Article.objects.get(id=article_id).like_set.count()
 
@@ -57,19 +72,21 @@ def like_articles_views(request):
 
 
 def favorite_articles_views(request):
-
     user_id = request.POST.get('user_id')
     article_id = request.POST.get('article_id')
 
-    self_favorite = Favorite.objects.filter(user_id=user_id, article_id=article_id).exists()
-    if self_favorite:
-        article_favorite = Favorite.objects.filter(user_id=user_id, article_id=article_id)
-        article_favorite.delete()
-        f_button_label = 'Favorite'
-    else:
-        article_favorite = Favorite(user_id=user_id, article_id=article_id)
-        article_favorite.save()
-        f_button_label = 'Unfavorire'
+    self_favorite = cache.get('self_favorite')
+    if not self_favorite:
+
+        self_favorite = Favorite.objects.filter(user_id=user_id, article_id=article_id).exists()
+        if self_favorite:
+            article_favorite = Favorite.objects.filter(user_id=user_id, article_id=article_id)
+            article_favorite.delete()
+            f_button_label = 'Favorite'
+        else:
+            article_favorite = Favorite(user_id=user_id, article_id=article_id)
+            article_favorite.save()
+            f_button_label = 'Unfavorire'
 
     return JsonResponse({'f_button_label':f_button_label})
 
@@ -83,9 +100,6 @@ def edit_article(request, pk):
             form.save()
     else:
         form = ArticleEditForm(instance=article)
-
-
-
 
     return render(request, 'edit_article.html', {'form':form, 'article':article})
